// src/app/constants/blog.enum.ts
export enum BlogStatus {
  DRAFT = "DRAFT",
  PUBLISHED = "PUBLISHED",
  ARCHIVED = "ARCHIVED",
}

export enum BlogCategory {
  CNC_GUIDE = "CNC_GUIDE",
  MACHINE_TIPS = "MACHINE_TIPS",
  BUSINESS_INSIGHT = "BUSINESS_INSIGHT",
  INDUSTRY_NEWS = "INDUSTRY_NEWS",
}

export enum HighlightBadge {
  NONE = "NONE",
  NEW = "NEW",
  FEATURED = "FEATURED",
  LIVE = "LIVE",
  ENDED = "ENDED",
}